requirements.in.txt

pydub
transformers
spacy
torchaudio
whisper-timestamped